requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
raven==6.9.0
blinker==1.4
celery==4.2.0
redis==2.10.6
pony==0.7.3
MySQL-python==1.2.5
Flask-Script==2.0.6
requests==2.19.1
Faker==0.8.16
apscheduler==3.5.1
gevent==1.3.0
gunicorn==19.9.0
bs4==0.0.1
xpinyin==0.5.5
APScheduler==3.5.1
pycrypto==2.6.1
lxml==4.2.3
ConcurrentLogHandler==0.9.1
selenium==3.14.0
ansible==2.6.3
